[Refactor] Transition save container to JSON string serialization
This commit refactors the game's save system to utilize JSON string serialization for actor-specific data, improving modularity, scalability, and reducing memory overhead compared to a large generic container.

Key changes include:
- Implemented JSON serialization and deserialization logic for `APlayerCharacter`, `AModularRobot`, and `ABaseHelperBot` specific save data. Each actor now serializes its unique data into a JSON string stored within `FSavedActorData::SerializedData`.
- Removed the `ESavedActorType` enum previously used for identifying spawnable classes during load. Class identification is now handled by storing the actor's explicit class name as a string (`FSavedActorData::ActorClassName`), which is then used as a key in `USpawnActorClassConfig`'s TMap to retrieve the correct `UClass` for spawning.
- Updated `USaveGameManager`'s `ApplyCoreData` to handle the new string-based class identification and JSON deserialization.
