+++ Content/CR4S/_Data/SaveLoad/BP_SpawnActorClassConfig.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:9debf9054cb19e080c1ea6b8e0baed56d08661a5a4024d61ec65d7d7bd918d24
-size 2300
+oid sha256:a01ae4c00e6f0d0747d5deb4f63fd8def6ec471a99929178ac1000851b1bff4e
+size 2327
